LOG-363. Refactor types to separate files
