+++ pkg/utils/deployment.go
@@ -1,0 +1,95 @@
+package utils
+
+import (
+	"fmt"
+
+	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
+	apps "k8s.io/api/apps/v1"
+	core "k8s.io/api/core/v1"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+//NewDeployment stubs an instance of a Deployment
+func NewDeployment(deploymentName string, namespace string, loggingComponent string, component string, podSpec core.PodSpec) *apps.Deployment {
+	return &apps.Deployment{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Deployment",
+			APIVersion: apps.SchemeGroupVersion.String(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      deploymentName,
+			Namespace: namespace,
+			Labels: map[string]string{
+				"provider":      "openshift",
+				"component":     component,
+				"logging-infra": loggingComponent,
+			},
+		},
+		Spec: apps.DeploymentSpec{
+			Replicas: GetInt32(1),
+			Selector: &metav1.LabelSelector{
+				MatchLabels: map[string]string{
+					"provider":      "openshift",
+					"component":     component,
+					"logging-infra": loggingComponent,
+				},
+			},
+			Template: core.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: deploymentName,
+					Labels: map[string]string{
+						"provider":      "openshift",
+						"component":     component,
+						"logging-infra": loggingComponent,
+					},
+				},
+				Spec: podSpec,
+			},
+			Strategy: apps.DeploymentStrategy{
+				Type: apps.RollingUpdateDeploymentStrategyType,
+				//RollingUpdate: {}
+			},
+		},
+	}
+}
+
+//GetDeploymentList returns a list for a give namespace and selector
+func GetDeploymentList(namespace, selector string) (*apps.DeploymentList, error) {
+	list := &apps.DeploymentList{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Deployment",
+			APIVersion: apps.SchemeGroupVersion.String(),
+		},
+	}
+
+	err := sdk.List(
+		namespace,
+		list,
+		sdk.WithListOptions(&metav1.ListOptions{
+			LabelSelector: selector,
+		}),
+	)
+
+	return list, err
+}
+
+//RemoveDeployment of given name and namespace
+func RemoveDeployment(namespace, deploymentName string) error {
+
+	deployment := NewDeployment(
+		deploymentName,
+		namespace,
+		deploymentName,
+		deploymentName,
+		core.PodSpec{},
+	)
+
+	err := sdk.Delete(deployment)
+	if err != nil && !errors.IsNotFound(err) {
+		return fmt.Errorf("Failure deleting %v deployment %v", deploymentName, err)
+	}
+
+	return nil
+}
