+++ pkg/utils/serviceaccount.go
@@ -1,0 +1,37 @@
+package utils
+
+import (
+	"fmt"
+
+	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
+	core "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+//NewServiceAccount stubs a new instance of ServiceAccount
+func NewServiceAccount(accountName string, namespace string) *core.ServiceAccount {
+	return &core.ServiceAccount{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "ServiceAccount",
+			APIVersion: core.SchemeGroupVersion.String(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      accountName,
+			Namespace: namespace,
+		},
+	}
+}
+
+//RemoveServiceAccount of given name and namespace
+func RemoveServiceAccount(namespace string, serviceAccountName string) error {
+
+	serviceAccount := NewServiceAccount(serviceAccountName, namespace)
+
+	err := sdk.Delete(serviceAccount)
+	if err != nil && !errors.IsNotFound(err) {
+		return fmt.Errorf("Failure deleting %v service account: %v", serviceAccountName, err)
+	}
+
+	return nil
+}
