+++ pkg/utils/utils.go
@@ -10,17 +10,9 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	logging "github.com/openshift/cluster-logging-operator/pkg/apis/logging/v1alpha1"
-
-	route "github.com/openshift/api/route/v1"
 	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
 	apps "k8s.io/api/apps/v1"
-	batch "k8s.io/api/batch/v1beta1"
-	core "k8s.io/api/core/v1"
-	rbac "k8s.io/api/rbac/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/util/retry"
 )
 
 const WORKING_DIR = "/tmp/_working_dir"
@@ -75,28 +67,6 @@ func GetComponentImage(component string) string {
 	}
 	logrus.Debugf("Setting component image for '%s' to: '%s'", component, imageTag)
 	return imageTag
-}
-
-// CreateClusterRole creates a cluser role or returns error
-func CreateClusterRole(name string, rules []rbac.PolicyRule, cluster *logging.ClusterLogging) (*rbac.ClusterRole, error) {
-	clusterRole := &rbac.ClusterRole{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "ClusterRole",
-			APIVersion: rbac.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Rules: rules,
-	}
-
-	AddOwnerRefToObject(clusterRole, AsOwner(cluster))
-
-	err := sdk.Create(clusterRole)
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return nil, fmt.Errorf("Failure creating '%s' clusterrole: %v", name, err)
-	}
-	return clusterRole, nil
 }
 
 func GetFileContents(filePath string) []byte {
@@ -161,25 +131,6 @@ func GetInt64(value int64) *int64 {
 	return &i
 }
 
-func GetDeploymentList(namespace string, selector string) (*apps.DeploymentList, error) {
-	list := &apps.DeploymentList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Deployment",
-			APIVersion: apps.SchemeGroupVersion.String(),
-		},
-	}
-
-	err := sdk.List(
-		namespace,
-		list,
-		sdk.WithListOptions(&metav1.ListOptions{
-			LabelSelector: selector,
-		}),
-	)
-
-	return list, err
-}
-
 func GetReplicaSetList(namespace string, selector string) (*apps.ReplicaSetList, error) {
 	list := &apps.ReplicaSetList{
 		TypeMeta: metav1.TypeMeta{
@@ -197,287 +148,4 @@ func GetReplicaSetList(namespace string, selector string) (*apps.ReplicaSetList,
 	)
 
 	return list, err
-}
-
-func GetPodList(namespace string, selector string) (*core.PodList, error) {
-	list := &core.PodList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: core.SchemeGroupVersion.String(),
-		},
-	}
-
-	err := sdk.List(
-		namespace,
-		list,
-		sdk.WithListOptions(&metav1.ListOptions{
-			LabelSelector: selector,
-		}),
-	)
-
-	return list, err
-}
-
-func GetCronJobList(namespace string, selector string) (*batch.CronJobList, error) {
-	list := &batch.CronJobList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "CronJob",
-			APIVersion: batch.SchemeGroupVersion.String(),
-		},
-	}
-
-	err := sdk.List(
-		namespace,
-		list,
-		sdk.WithListOptions(&metav1.ListOptions{
-			LabelSelector: selector,
-		}),
-	)
-
-	return list, err
-}
-
-func GetDaemonSetList(namespace string, selector string) (*apps.DaemonSetList, error) {
-	list := &apps.DaemonSetList{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "DaemonSet",
-			APIVersion: apps.SchemeGroupVersion.String(),
-		},
-	}
-
-	err := sdk.List(
-		namespace,
-		list,
-		sdk.WithListOptions(&metav1.ListOptions{
-			LabelSelector: selector,
-		}),
-	)
-
-	return list, err
-}
-
-func GetRouteURL(routeName, namespace string) (string, error) {
-
-	foundRoute := &route.Route{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Route",
-			APIVersion: route.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      routeName,
-			Namespace: namespace,
-			Labels: map[string]string{
-				"component":     "support",
-				"logging-infra": "support",
-				"provider":      "openshift",
-			},
-		},
-	}
-
-	if err := sdk.Get(foundRoute); err != nil {
-		if !errors.IsNotFound(err) {
-			logrus.Errorf("Failed to check for ClusterLogging object: %v", err)
-		}
-		return "", err
-	}
-
-	return fmt.Sprintf("%s%s", "https://", foundRoute.Spec.Host), nil
-}
-
-func CreateOrUpdateSecret(secret *core.Secret) (err error) {
-	err = sdk.Create(secret)
-	if err != nil {
-		if !errors.IsAlreadyExists(err) {
-			return fmt.Errorf("Failure constructing %v secret: %v", secret.Name, err)
-		}
-
-		current := secret.DeepCopy()
-		retryErr := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-			if err = sdk.Get(current); err != nil {
-				if errors.IsNotFound(err) {
-					// the object doesn't exist -- it was likely culled
-					// recreate it on the next time through if necessary
-					return nil
-				}
-				return fmt.Errorf("Failed to get %v secret: %v", secret.Name, err)
-			}
-
-			current.Data = secret.Data
-			if err = sdk.Update(current); err != nil {
-				return err
-			}
-			return nil
-		})
-		if retryErr != nil {
-			return retryErr
-		}
-	}
-
-	return nil
-}
-
-func RemoveServiceAccount(namespace string, serviceAccountName string) error {
-
-	serviceAccount := ServiceAccount(serviceAccountName, namespace)
-
-	err := sdk.Delete(serviceAccount)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v service account: %v", serviceAccountName, err)
-	}
-
-	return nil
-}
-
-func RemoveConfigMap(namespace string, configmapName string) error {
-
-	configMap := ConfigMap(
-		configmapName,
-		namespace,
-		map[string]string{},
-	)
-
-	err := sdk.Delete(configMap)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v configmap: %v", configmapName, err)
-	}
-
-	return nil
-}
-
-func RemoveSecret(namespace string, secretName string) error {
-
-	secret := Secret(
-		secretName,
-		namespace,
-		map[string][]byte{},
-	)
-
-	err := sdk.Delete(secret)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v secret: %v", secretName, err)
-	}
-
-	return nil
-}
-
-func RemoveRoute(namespace string, routeName string) error {
-
-	route := Route(
-		routeName,
-		namespace,
-		routeName,
-		"",
-	)
-
-	err := sdk.Delete(route)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v route %v", routeName, err)
-	}
-
-	return nil
-}
-
-func RemoveService(namespace string, serviceName string) error {
-
-	service := Service(
-		serviceName,
-		namespace,
-		serviceName,
-		[]core.ServicePort{},
-	)
-
-	err := sdk.Delete(service)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v service %v", serviceName, err)
-	}
-
-	return nil
-}
-
-func RemoveOAuthClient(namespace string, clientName string) error {
-
-	oauthClient := OAuthClient(
-		clientName,
-		namespace,
-		"",
-		[]string{},
-		[]string{},
-	)
-
-	err := sdk.Delete(oauthClient)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v oauthclient %v", clientName, err)
-	}
-
-	return nil
-}
-
-func RemoveDaemonset(namespace string, daemonsetName string) error {
-
-	daemonset := DaemonSet(
-		daemonsetName,
-		namespace,
-		daemonsetName,
-		daemonsetName,
-		core.PodSpec{},
-	)
-
-	err := sdk.Delete(daemonset)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v daemonset %v", daemonsetName, err)
-	}
-
-	return nil
-}
-
-func RemoveDeployment(namespace string, deploymentName string) error {
-
-	deployment := Deployment(
-		deploymentName,
-		namespace,
-		deploymentName,
-		deploymentName,
-		core.PodSpec{},
-	)
-
-	err := sdk.Delete(deployment)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v deployment %v", deploymentName, err)
-	}
-
-	return nil
-}
-
-func RemoveCronJob(namespace string, cronjobName string) error {
-
-	cronjob := CronJob(
-		cronjobName,
-		namespace,
-		cronjobName,
-		cronjobName,
-		batch.CronJobSpec{},
-	)
-
-	err := sdk.Delete(cronjob)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v cronjob %v", cronjobName, err)
-	}
-
-	return nil
-}
-
-func RemovePriorityClass(priorityclassName string) error {
-	collectionPriorityClass := PriorityClass(
-		priorityclassName,
-		1000000,
-		false,
-		"This priority class is for the Cluster-Logging Collector",
-	)
-
-	err := sdk.Delete(collectionPriorityClass)
-	if err != nil && !errors.IsNotFound(err) {
-		return fmt.Errorf("Failure deleting %v priority class %v", priorityclassName, err)
-	}
-
-	return nil
 }
